requirements.txt
asgiref==3.9.2
psycopg2-binary==2.9.10
attrs==25.4.0
diploma-frontend @ file:///Users/aknursaripzan/PycharmProjects/python_django_diploma/diploma-frontend/dist/diploma-frontend-0.6.tar.gz#sha256=3a6b1203724b256946919b4d65b362b005ae6a9e1e93d8d765e96aa6a92764ac
Django==5.2.6
django-cleanup==9.0.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
factory_boy==3.3.3
Faker==39.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
